Add files via upload
Matlab scripts to generate Shock response spectrum of a payload deployed using a chosen commercially available separation mechanisms. SRS contain the different options for the mechanisms. Running this file will save the data in external files. Equation_Solve contains the governing equation to be solved. Diff_Mechanism_plots contain script to plot all the  data stored for each of the mechanism run in a single figure.